+++ apps/projects/ci-apps/cadastros/src/lib/cadastros/master-detail/master-detail.component.ts
@@ -5,7 +5,7 @@ import { MatIconModule } from "@angular/material/icon";
 import { MatToolbarModule } from "@angular/material/toolbar";
 import { ActivatedRoute, RouterModule } from "@angular/router";
 import { GridModule, IColumnOption, IDataGridOptions, WindowModule, WindowService } from "@ci/components";
-import { CoreModule, DaoBuilder, DaoService } from "@ci/core";
+import { CoreModule, DaoBuilder, DaoService, IHaveSync, ISchema } from "@ci/core";
 import { Application, getServiceAsSchema } from "@ci/portal-api";
 import { EditarComponent } from "./editar/editar.component";
 import { lastValueFrom } from "rxjs";
@@ -43,9 +43,11 @@ export class MasterDetailComponent<T> implements OnInit, AfterViewInit {
 
     }
     source?: T[] = [{} as any];
+    properties?: ISchema;
     async loadGrid() {
         if (!!this.schemaName) {
             const properties = await (await this.daoBuilder.getSchema(this.schemaName)).properties
+            this.properties = properties;
             this.gridOptions = {
                 columns: [
                     ...Object.keys(properties || {}).map(property => {
@@ -100,7 +102,7 @@ export class MasterDetailComponent<T> implements OnInit, AfterViewInit {
     }
     async search() {
         if (this.service && this.service.getList)
-            this.source = await this.daos?.read(await lastValueFrom(this.service.getList()));
+            this.source = await this.daos?.read(await lastValueFrom(this.service.getList()), this.schemaName);
     }
     async editar(data: T, event?: Event) {
         return await this.window.open(EditarComponent,
@@ -110,7 +112,8 @@ export class MasterDetailComponent<T> implements OnInit, AfterViewInit {
     }
     async createNew() {
         let new_instance: T = {} as T;
-        /* const data =  */await this.editar(new_instance);
+        let new_instance_result: any = await lastValueFrom((this.service as IHaveSync<any>).sync({ body: { data: new_instance } }))
+        /* const data =  */await this.editar(new_instance_result);
         /* if (!!data?.internalId || !!data?.id) // TODO: revisar esta regra
             this.source = [data, ...this.source || []];
         else {