+++ apps/projects/ci-apps/low-code/src/lib/ci-application/home/blocks-maker/block-code/block-code.component.ts
@@ -15,6 +15,5 @@ export class CodeBlockComponent {
     @Input() description?: string;
     @Input() inputs?: any[];
     @Input() outputs?: any[];
-
     constructor() { }
 }