+++ apps/projects/ci-apps/low-code/src/lib/ci-application/home/blocks-maker/blocks-maker.component.ts
@@ -19,4 +19,7 @@ import { CoreModule } from '@ci/core';
 })
 export class BlocksMakerComponent {
   protected blocos?: any[];
+  async createBlock() {
+    
+  }
 }