+++ apps/projects/ci/components/src/lib/footer/footer.component.ts
@@ -1,5 +1,4 @@
-import { Component, OnInit } from "@angular/core";
-
+import { Component } from "@angular/core";
 @Component({
     selector: 'x-footer',
     standalone: false,
@@ -7,17 +6,13 @@ import { Component, OnInit } from "@angular/core";
     template: `
     <footer>
         <div class="assinatura" style="flex:auto; display:flex; flex-direction: row;">
-            Desenvolvido por <a href="https://ci.dev.br" target="_blank">ci.dev.br</a> 2017 - {{ano}}
-            <span style="flex:auto"></span>
+            Produzido em <a href="https://ci.dev.br" target="_blank">ci.dev.br</a> 2017 - {{ano}}
+            <a por="/credits-of-autrhors"></a><span style="flex:auto"></span>
             <a href="/termos-de-uso">Termos de Uso e Privacidade</a>
         </div>
     </footer>
     `
 })
-export class Footer implements OnInit {
+export class Footer {
     protected ano = new Date().getFullYear();
-    constructor() { }
-    async ngOnInit() {
-
-    }
 }