+++ apps/projects/ci/components/src/lib/models/i-data-grid-options.ts
@@ -1,3 +1,4 @@
+import { IContextMenuOf } from "@ci/components/context-menu";
 import { IColumnOption } from "./i-column-options";
 import { ILoadListOptions } from "./i-load-list-options";
 import { IPivot } from "./i-pivot";
@@ -6,4 +7,5 @@ export interface IDataGridOptions<T> {
     columns: IColumnOption<T>[];
     group?: IPivot<T>;
     loadList?: (options: ILoadListOptions<T>) => T[] | Promise<T[]>;
+    contextmenu?: IContextMenuOf<T>[];
 }