+++ apps/projects/ci/portal-api/src/lib/models/obter-lista-endereco.ts
@@ -1,6 +1,8 @@
 /* tslint:disable */
 /* eslint-disable */
 export interface ObterListaEndereco {
+  relations: {
+};
   skip: number;
   take: number;
   where: {