+++ server/libs/cadastro/src/model/endereco.entity.ts
@@ -39,7 +39,6 @@ export class Endereco extends FullAuditedEntity {
      */
     @ApiProperty({
         title: 'Número',
-
         nullable: true,
         required: false
     })
@@ -52,7 +51,6 @@ export class Endereco extends FullAuditedEntity {
      */
     @ApiProperty({
         title: 'Complemento',
-
         nullable: true,
         required: false
     })
@@ -66,7 +64,7 @@ export class Endereco extends FullAuditedEntity {
     @ApiProperty({
         title: 'Município',
         nullable: true,
-        required: false
+        required: false,
     })
     @Column({
         nullable: true